Add dummy user to represent deleted JIRA users
To stop the migration failing in the case where a deleted JIRA user is
associated with an issue/comment/attachment, I added a 'ghost' user.
The ghost user is used to represent all deleted JIRA users.

This ghost user is only migrated to Redmine if it is needed.
It is locked (disabled), and it has the login name 'deleted-user'.